Original, full data for Sobotik 1998
Original data of relevés for the (synthetic) table in Sobotik 1993 - see
pdf in this directory.
The data here brings four relevés per plot (16), made between 1993 -
1996 (four relevés); from this data the synthetic table was derived. The
last digit in this data/plot id is the number of the relevé, the first,
respective first two digit(s) are the plots as in the synthetic table,
e.g. Plot 1 from the synthetic table is derived from relevés
11,12,13,14, for plot 16 the relevés 161, 162, 163, 164 and so on.